move paste to world
